+++ src/pages/UserInfoPage.tsx
@@ -1,9 +1,11 @@
-type UserInfoProps = {
+import React from 'react';
+
+interface UserInfoPageProps {
   nickname: string;
   onLogout: () => void;
-};
+}
 
-const UserInfo = ({ nickname, onLogout }: UserInfoProps) => {
+const UserInfoPage: React.FC<UserInfoPageProps> = ({ nickname, onLogout }) => {
   return (
     <div className="h-full w-full flex flex-col justify-center items-center bg-white">
       <h1 className="text-2xl font-bold mb-4">{nickname}</h1>
@@ -17,4 +19,4 @@ const UserInfo = ({ nickname, onLogout }: UserInfoProps) => {
   );
 };
 
-export default UserInfo;
+export default UserInfoPage;
